CAR PRICE
You are required to model the prices of cars with the available independent variables. It will be
used by management to understand how exactly the prices vary with the independent variables.
They can accordingly manipulate the design of the cars, the business strategy etc... to meet
certain price levels. Further, the model will be good for management to understand the pricing
dynamics of the new market.